requirements.txt
PyYAML>=5.1.1
Flask
PyMySQL
pytz
pytest
Crypto